pyproject.toml
[tool.poetry]
name = "payment-api"
version = "0.1.0"
description = ""
authors = ["Alex Kosarevsky <alexkosarevsky@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.7"
djangorestframework = "^3.15.2"
django-filter = "^25.1"
gunicorn = "^23.0.0"
psycopg2 = "^2.9.10"
pytest = "^8.3.5"
ruff = "^0.11.1"
pytest-django = "^4.10.0"
pytest-mock = "^3.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
